Reorganise project structure and add packages
Current structure is not very organised after adding new classes.

New packages are added and classes are moved into packages.

* dude.ui package added
* dude.note package added
