+++ src/main/java/dude/Main.java
@@ -2,6 +2,7 @@ package dude;
 
 import java.io.IOException;
 
+import dude.ui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
