+++ src/main/java/dude/command/AddToDoCommand.java
@@ -2,10 +2,10 @@ package dude.command;
 
 import java.io.IOException;
 
-import dude.NoteList;
+import dude.note.NoteList;
 import dude.Storage;
-import dude.TaskList;
-import dude.Ui;
+import dude.task.TaskList;
+import dude.ui.Ui;
 import dude.task.ToDo;
 
 /**
