+++ src/main/java/dude/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package dude;
+package dude.ui;
 
 import java.io.IOException;
 import java.util.Collections;
