+++ src/main/java/dude/ui/MainWindow.java
@@ -1,5 +1,6 @@
-package dude;
+package dude.ui;
 
+import dude.Dude;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
